refactor tests: add utils helpers
